+++ CHANGELOG.rst
@@ -17,6 +17,16 @@ Changelog
     filename including its path and suffix. (In turn, the parameter
     ``store_simulation`` was removed.)
 
+- ``simulations.Simulation``:
+
+  - Warns if the gradient is called, but ``receiver_interpolation`` is not
+    ``'linear'``.
+
+- Various:
+
+  - All emg3d-warnings (not solver warnings) are now set to ``'always'``, and
+    corresponding print statements were removed.
+
 
 v1.1.0: Adjoint-fix for electric receivers
 ------------------------------------------